fix: move modules into place where used
