+++ src/components/Item/VisualizationItem/getGridItemElement.js
@@ -1,4 +1,4 @@
-import { getGridItemDomElementClassName } from './getGridItemDomElementClassName'
+import { getGridItemDomElementClassName } from '../../../modules/getGridItemDomElementClassName'
 
 export const getGridItemElement = itemId =>
     document.querySelector(`.${getGridItemDomElementClassName(itemId)}`)
