+++ src/components/Item/VisualizationItem/isElementFullscreen.js
@@ -1,4 +1,4 @@
-import { getGridItemDomElementClassName } from './getGridItemDomElementClassName'
+import { getGridItemDomElementClassName } from '../../../modules/getGridItemDomElementClassName'
 
 export const isElementFullscreen = itemId => {
     const fullscreenElement =
